third_party/blink/web_tests/platform/linux-chrome/external/wpt/editing/other/typing-around-link-element-at-non-collapsed-selection.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Inserting "XY" after deleting first character of a link in <b> (Direct typing) in <p><b><a href="about:blank">[z]abc</a></b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">zabc</a></b></p>" not in array ["<p><b><a href=\"about:blank\">XYabc</a></b></p>", "<p><b><a href=\"about:blank\">XYabc</a></b><br></p>"]
[FAIL] Inserting "XY" after deleting last character in a non-collapsed range of a link in <b> (Direct typing) in <p><b><a href="about:blank">abc[d]</a></b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">abcd</a></b></p>" not in array ["<p><b><a href=\"about:blank\">abcXY</a></b></p>", "<p><b><a href=\"about:blank\">abcXY</a></b><br></p>"]
[FAIL] Inserting "XY" after deleting text after middle of a link in <b> (Direct typing) in <p><b><a href="about:blank">ab[cd</a>de]f</b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">abcd</a>def</b></p>" not in array ["<p><b><a href=\"about:blank\">ab</a>XYf</b></p>", "<p><b><a href=\"about:blank\">ab</a>XYf</b><br></p>"]
[FAIL] Inserting "XY" after deleting text before middle of a link in <b> (Direct typing) in <p><b>a[bc<a href="about:blank">de]f</a></b></p>
  assert_in_array: value "<p><b>YXabc<a href=\"about:blank\">def</a></b></p>" not in array ["<p><b>aXY<a href=\"about:blank\">f</a></b></p>", "<p><b>aXY<a href=\"about:blank\">f</a></b><br></p>"]
[FAIL] Inserting "XY" after deleting first character of a link in <b> (Backspace) in <p><b><a href="about:blank">[z]abc</a></b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">zabc</a></b></p>" not in array ["<p><b>XY<a href=\"about:blank\">abc</a></b></p>", "<p><b>XY<a href=\"about:blank\">abc</a></b><br></p>"]
[FAIL] Inserting "XY" after deleting last character in a non-collapsed range of a link in <b> (Backspace) in <p><b><a href="about:blank">abc[d]</a></b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">abcd</a></b></p>" not in array ["<p><b><a href=\"about:blank\">abc</a>XY</b></p>", "<p><b><a href=\"about:blank\">abc</a>XY</b><br></p>"]
[FAIL] Inserting "XY" after deleting text after middle of a link in <b> (Backspace) in <p><b><a href="about:blank">ab[cd</a>de]f</b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">abcd</a>def</b></p>" not in array ["<p><b><a href=\"about:blank\">ab</a>XYf</b></p>", "<p><b><a href=\"about:blank\">ab</a>XYf</b><br></p>"]
[FAIL] Inserting "XY" after deleting text before middle of a link in <b> (Backspace) in <p><b>a[bc<a href="about:blank">de]f</a></b></p>
  assert_in_array: value "<p><b>YXabc<a href=\"about:blank\">def</a></b></p>" not in array ["<p><b>aXY<a href=\"about:blank\">f</a></b></p>", "<p><b>aXY<a href=\"about:blank\">f</a></b><br></p>"]
[FAIL] Inserting "XY" after deleting first character of a link in <b> (Delete) in <p><b><a href="about:blank">[z]abc</a></b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">abc</a></b></p>" not in array ["<p><b>XY<a href=\"about:blank\">abc</a></b></p>", "<p><b>XY<a href=\"about:blank\">abc</a></b><br></p>"]
[FAIL] Inserting "XY" after deleting last character in a non-collapsed range of a link in <b> (Delete) in <p><b><a href="about:blank">abc[d]</a></b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">bcd</a></b></p>" not in array ["<p><b><a href=\"about:blank\">abc</a>XY</b></p>", "<p><b><a href=\"about:blank\">abc</a>XY</b><br></p>"]
[FAIL] Inserting "XY" after deleting text after middle of a link in <b> (Delete) in <p><b><a href="about:blank">ab[cd</a>de]f</b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">bcd</a>def</b></p>" not in array ["<p><b><a href=\"about:blank\">ab</a>XYf</b></p>", "<p><b><a href=\"about:blank\">ab</a>XYf</b><br></p>"]
[FAIL] Inserting "XY" after deleting text before middle of a link in <b> (Delete) in <p><b>a[bc<a href="about:blank">de]f</a></b></p>
  assert_in_array: value "<p><b>YXbc<a href=\"about:blank\">def</a></b></p>" not in array ["<p><b>aXY<a href=\"about:blank\">f</a></b></p>", "<p><b>aXY<a href=\"about:blank\">f</a></b><br></p>"]
[FAIL] Inserting "XY" after deleting first character of a link in <b> (execCommand("delete")) in <p><b><a href="about:blank">[z]abc</a></b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">abc</a></b></p>" not in array ["<p><b>XY<a href=\"about:blank\">abc</a></b></p>", "<p><b>XY<a href=\"about:blank\">abc</a></b><br></p>"]
[FAIL] Inserting "XY" after deleting last character in a non-collapsed range of a link in <b> (execCommand("delete")) in <p><b><a href="about:blank">abc[d]</a></b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">abc</a></b></p>" not in array ["<p><b><a href=\"about:blank\">abc</a>XY</b></p>", "<p><b><a href=\"about:blank\">abc</a>XY</b><br></p>"]
[FAIL] Inserting "XY" after deleting text after middle of a link in <b> (execCommand("delete")) in <p><b><a href="about:blank">ab[cd</a>de]f</b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">ab</a>f</b></p>" not in array ["<p><b><a href=\"about:blank\">ab</a>XYf</b></p>", "<p><b><a href=\"about:blank\">ab</a>XYf</b><br></p>"]
[FAIL] Inserting "XY" after deleting text before middle of a link in <b> (execCommand("delete")) in <p><b>a[bc<a href="about:blank">de]f</a></b></p>
  assert_in_array: value "<p><b>YXa<a href=\"about:blank\">f</a></b></p>" not in array ["<p><b>aXY<a href=\"about:blank\">f</a></b></p>", "<p><b>aXY<a href=\"about:blank\">f</a></b><br></p>"]
[FAIL] Inserting "XY" after deleting first character of a link in <b> (execCommand("forwarddelete")) in <p><b><a href="about:blank">[z]abc</a></b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">abc</a></b></p>" not in array ["<p><b>XY<a href=\"about:blank\">abc</a></b></p>", "<p><b>XY<a href=\"about:blank\">abc</a></b><br></p>"]
[FAIL] Inserting "XY" after deleting last character in a non-collapsed range of a link in <b> (execCommand("forwarddelete")) in <p><b><a href="about:blank">abc[d]</a></b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">abc</a></b></p>" not in array ["<p><b><a href=\"about:blank\">abc</a>XY</b></p>", "<p><b><a href=\"about:blank\">abc</a>XY</b><br></p>"]
[FAIL] Inserting "XY" after deleting text after middle of a link in <b> (execCommand("forwarddelete")) in <p><b><a href="about:blank">ab[cd</a>de]f</b></p>
  assert_in_array: value "<p><b>YX<a href=\"about:blank\">ab</a>f</b></p>" not in array ["<p><b><a href=\"about:blank\">ab</a>XYf</b></p>", "<p><b><a href=\"about:blank\">ab</a>XYf</b><br></p>"]
[FAIL] Inserting "XY" after deleting text before middle of a link in <b> (execCommand("forwarddelete")) in <p><b>a[bc<a href="about:blank">de]f</a></b></p>
  assert_in_array: value "<p><b>YXa<a href=\"about:blank\">f</a></b></p>" not in array ["<p><b>aXY<a href=\"about:blank\">f</a></b></p>", "<p><b>aXY<a href=\"about:blank\">f</a></b><br></p>"]
Harness: the test ran to completion.
